Optimize miscellaneous scripts

+++ catscripts/Entertainment/Cartoons/Avatar_The_Last_Airbender/Avatar.sh
@@ -3,15 +3,29 @@
 KEYWORDS_AIRBENDER="\bAang|Katara|\bToph(|s)\b|(air|fire|water|earth)(| )(bender|tribe|people)|avatar cycle|\bSokka|\bZuko|\bAppa(|(|')s)\b|\bIroh(|s)\b|\bMomo\b|\bAzula\b|\bOzal\b"
 KEYWORDS_AIRBENDER_EXCLUDE="Momo (Sohma|Hinamori|Adachi)|Bleach"
 
-egrep -i "$KEYWORDS_AIRBENDER" newpages.txt | egrep -iv "$KEYWORDS_AIRBENDER_EXCLUDE" >> TheLastAirbender.txt
-
-AVATAR=`stat --print=%s TheLastAirbender.txt`
-
-if [ $AVATAR -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="TheLastAirbender.txt"
-  export CATNAME="Avatar: The Last Airbender"
-  $CATEGORIZE
-fi
 
-rm TheLastAirbender.txt
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Avatar: The Last Airbender\n"
+  fi
+
+  AVATAR=`egrep -i "$KEYWORDS_AIRBENDER" newpages.txt | egrep -iv "$KEYWORDS_AIRBENDER_EXCLUDE"`
+
+  if [ "$AVATAR" != "" ];
+  then
+    printf "$AVATAR" > TheLastAirbender.txt
+    export CATFILE="TheLastAirbender.txt"
+    export CATNAME="Avatar: The Last Airbender"
+    $CATEGORIZE
+    rm TheLastAirbender.txt
+    unset AVATAR
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Avatar: The Last Airbender\n"
+  fi
+
+fi
