+++ catscripts/Entertainment/Franchises/Pokémon/Pokémon_species/X/X.sh
@@ -1,13 +1,30 @@
 #!/bin/bash
-egrep -i 'Xatu' newpages.txt >> Xatu.txt
 
-Xatu=`stat --print=%s Xatu.txt`
+KEYWORDS_XATU="Xatu"
 
-if [ $Xatu -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Xatu.txt"
-  export CATNAME="Xatu"
-  $CATEGORIZE
-fi
 
-rm Xatu.txt
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Pokemon X scripts\n"
+  fi
+
+  XATU=`egrep -i "$KEYWORDS_XATU" newpages.txt`
+
+  if [ "$XATU" != "" ];
+  then
+    printf "$XATU" > Xatu.txt
+    export CATFILE="Xatu.txt"
+    export CATNAME="Xatu"
+    $CATEGORIZE
+    rm Xatu.txt
+    unset XATU
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Pokemon X scripts\n"
+  fi
+
+fi
