+++ catscripts/Entertainment/Franchises/Pokémon/Pokémon_species/Y/Y.sh
@@ -1,34 +1,55 @@
 #!/bin/bash
-egrep -i 'Yamask' newpages.txt >> Yamask.txt
-egrep -i '\bYanma\b' newpages.txt >> Yanma.txt
-egrep -i 'Yanmega' newpages.txt >> Yanmega.txt
 
-Yamask=`stat --print=%s Yamask.txt`
-Yanma=`stat --print=%s Yanma.txt`
-Yanmega=`stat --print=%s Yanmega.txt`
+KEYWORDS_YAMASK="Yamask"
+KEYWORDS_YANMA="\bYanma"
+KEYWORDS_YANMEGA="Yanmega"
 
-if [ $Yamask -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Yamask.txt"
-  export CATNAME="Yamask"
-  $CATEGORIZE
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Pokemon Y script\n"
+  fi
+
+  YAMASK=`egrep -i "$KEYWORDS_YAMASK" newpages.txt`
+  YANMA=`egrep -i "$KEYWORDS_YANMA" newpages.txt`
+  YANMEGA=`egrep -i "$KEYWORDS_YANMEGA" newpages.txt`
+
+  if [ "$YAMASK" != "" ];
+  then
+    printf "$YAMASK" > Yamask.txt
+    export CATFILE="Yamask.txt"
+    export CATNAME="Yamask"
+    $CATEGORIZE
+    rm Yamask.txt
+    unset YAMASK
+  fi
+
+  if [ "$YANMA" != "" ];
+  then
+    printf "$YANMA" > Yanma.txt
+    export CATFILE="Yanma.txt"
+    export CATNAME="Yanma"
+    $CATEGORIZE
+    rm Yanma.txt
+    unset YANMA
+  fi
+
+  if [ "$YANMEGA" != "" ];
+  then
+    printf "$YANMEGA" > Yanmega.txt
+    export CATFILE="Yanmega.txt"
+    export CATNAME="Yanmega"
+    $CATEGORIZE
+    rm Yanmega.txt
+    unset YANMEGA
+  fi
+
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Pokemon Y script\n"
+  fi
+
 fi
-
-if [ $Yanma -ne 0 ];
-then
-  export CATFILE="Yanma.txt"
-  export CATNAME="Yanma"
-  $CATEGORIZE
-fi
-
-if [ $Yanmega -ne 0 ];
-then
-  export CATFILE="Yanmega.txt"
-  export CATNAME="Yanmega"
-  $CATEGORIZE
-fi
-
-
-rm Yamask.txt
-rm Yanma.txt
-rm Yanmega.txt
