+++ catscripts/Government/Countries/Laos/Laos.sh
@@ -1,14 +1,30 @@
 #!/bin/bash
 
-egrep -i 'Laos' newpages.txt >> Laos.txt
+KEYWORDS_LAOS="Laos"
 
-LAOS=`stat --print=%s Laos.txt`
-
-if [ $LAOS -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Laos.txt"
-  export CATNAME="Laos"
-  $CATEGORIZE
-fi
 
-rm Laos.txt
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Laos\n"
+  fi
+
+  LAOS=`egrep -i "$KEYWORDS_LAOS" newpages.txt`
+
+  if [ "$LAOS" != "" ];
+  then
+    printf "$LAOS" > Laos.txt
+    export CATFILE="Laos.txt"
+    export CATNAME="Laos"
+    $CATEGORIZE
+    rm Laos.txt
+    unset LAOS
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Laos\n"
+  fi
+
+fi
