+++ catscripts/Government/Countries/United_States/US_states/Maine/Maine.sh
@@ -1,14 +1,30 @@
 #!/bin/bash
 
-egrep -i '\bMaine' newpages.txt >> Maine.txt
+KEYWORDS_MAINE="\bMaine"
 
-MAINE=`stat --print=%s Maine.txt`
-
-if [ $MAINE -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Maine.txt"
-  export CATNAME="Maine"
-  $CATEGORIZE
-fi
 
-rm Maine.txt
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Maine\n"
+  fi
+
+  MAINE=`egrep -i "$KEYWORDS_MAINE" newpages.txt`
+
+  if [ "$MAINE" != "" ];
+  then
+    printf "$MAINE" > Maine.txt
+    export CATFILE="Maine.txt"
+    export CATNAME="Maine"
+    $CATEGORIZE
+    rm Maine.txt
+    unset MAINE
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Maine\n"
+  fi
+
+fi
